@@ -24,19 +24,22 @@
     "node-uuid": "1.4.8",
     "prop-types": "15.6.2",
     "pug": "2.0.0-rc.1",
-    "react": "15.0.1",
-    "react-dom": "15.0.1",
+    "react": "15.3.0",
+    "react-dom": "15.3.0",
     "react-fa": "4.0.1",
     "uuid": "2.0.2"
   },
   "devDependencies": {
+    "@babel/core": "^7.0.0",
+    "@babel/plugin-proposal-class-properties": "^7.4.4",
+    "@babel/plugin-transform-runtime": "^7.4.4",
+    "@babel/preset-env": "^7.0.0",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/runtime": "^7.4.4",
     "autoprefixer": "6.3.6",
-    "babel": "6.5.2",
-    "babel-core": "6.24.0",
-    "babel-eslint": "7.2.3",
-    "babel-loader": "6.2.4",
-    "babel-preset-es2015": "6.6.0",
-    "babel-preset-react": "6.5.0",
+    "babel-core": "^7.0.0-bridge.0",
+    "babel-eslint": "^9.0.0",
+    "babel-loader": "7.1.1",
     "css-loader": "0.23.1",
     "del": "2.2.0",
     "extract-text-webpack-plugin": "1.0.1",
