@@ -1,4 +1,4 @@
-const Dispatcher = require('flux').Dispatcher;
+import { Dispatcher } from 'flux';
 
 // Export a singleton instance of dispatcher, for use by the entire app
 export default new Dispatcher();
