@@ -1,22 +1,21 @@
 import React from 'react';
-import { searchImage } from '../../utils/api/create-api';
+import PropTypes from 'prop-types';
+import imageService from '../../services/image-service';
 
 import './search-box-view.css';
 
 const SEARCH_BOX_SIZE = 40;
 
-export default React.createClass({
-    displayName: 'search-box-view',
-
-    propTypes: {
-        placeholder: React.PropTypes.string
-    },
+export default class SearchBoxView extends React.Component {
+    constructor(props, context) {
+        super(props, context);
+    }
 
     onKeyPressed(event) {
         if (event.key === 'Enter') {
-            searchImage(event.target.value);
+            imageService.searchImage(event.target.value);
         }
-    },
+    }
 
     render() {
         return (
@@ -31,4 +30,8 @@ export default React.createClass({
             </div>
         );
     }
-});
+}
+
+SearchBoxView.propTypes = {
+    placeholder: PropTypes.string
+}
