@@ -1,85 +1,20 @@
 import React from 'react';
-import Icon from 'react-fa';
-import { v4 } from 'node-uuid'; // Yes this works in the browser too
-import { createRemoveFromCartAction } from '../../actions/cart-actions';
+import PropTypes from 'prop-types';
+
 import GnomeListView from '../../views/gnome-list-view/gnome-list-view';
 import { createExpandItemAction } from '../../actions/cart-actions';
+import AuthService from '../../services/auth-service';
+
+import ItemRow from './item-row';
 
 import './normal-cart-view.css';
 
-// Temporary until we implement actual auth
-if (!localStorage.token) {
-    localStorage.token = v4();
-}
-const token = localStorage.token;
 
-const ItemRow = React.createClass({
-
-    displayName: 'item-row',
-
-    propTypes: {
-        item: React.PropTypes.object
-    },
-
-    getInitialState() {
-        return {
-            quantity: 1
-        };
-    },
-
-    onQuantityChanged(e) {
-        this.setState({ quantity: e.target.value });
-    },
-
-    onItemRemoved() {
-        createRemoveFromCartAction(this.props.item);
-    },
-
-    render() {
-        const item = this.props.item;
-        return (
-            <div key={item.id} className="gs-cartview-normal-leftpane-row">
-                <div className="gs-cartview-normal-leftpane-row-product">
-                    <img src={item.image} />
-                    <div>
-                        <div>
-                            <span className="gs-cartview-normal-leftpane-row-product-name">{item.name}</span>
-                        </div>
-                        <div>
-                            <span className="gs-cartview-normal-leftpane-row-product-by"> by </span>
-                            <span className="gs-cartview-normal-leftpane-row-product-author">{item.author}</span>
-                        </div>
-                        <div className="gs-cartview-normal-leftpane-row-product-size">{item.size.width} x {item.size.height}</div>
-                    </div>
-                </div>
-                <div className="gs-cartview-normal-leftpane-row-total">$2.99</div>
-                <div className="gs-cartview-normal-leftpane-row-quantity">
-                    <select name={`checkout-items[${item.id}][quantity]`}
-                            value={this.state.quantity}
-                            onChange={this.onQuantityChanged}
-                            data-item-id={item.id}>
-                        <option value="1">1</option>
-                        <option value="2">2</option>
-                        <option value="3">3</option>
-                    </select>
-                </div>
-                <Icon name="close" className="gs-cartview-normal-leftpane-close" onClick={this.onItemRemoved} />
-                <input type="hidden" value={item.id} name={`checkout-items[${item.id}][id]`} />
-            </div>
-        );
+export default class NormalCartView extends React.Component {
+    constructor(props, context) {
+        super(props, context);
+        this.token = AuthService.getToken();
     }
-});
-
-
-
-export default React.createClass({
-
-    displayName: 'normal-cart-view',
-    
-    propTypes: {
-        items: React.PropTypes.arrayOf(React.PropTypes.object).isRequired,
-        recommendations: React.PropTypes.arrayOf(React.PropTypes.object)
-    },
 
     render() {
         let items = this.props.items;
@@ -104,7 +39,7 @@ export default React.createClass({
                             <div className="gs-cartview-normal-leftpane-header-quantity">Quantity</div>
                         </div>
                         {itemRows}
-                        <input type="hidden" name="token" value={token} />
+                        <input type="hidden" name="token" value={this.token} />
                     </div>
 
                     <div className="gs-cartview-normal-rightpane">
@@ -127,4 +62,10 @@ export default React.createClass({
             </div>
         );
     }
-});
+}
+
+NormalCartView.propTypes = {
+    items: PropTypes.arrayOf(PropTypes.object).isRequired,
+    recommendations: PropTypes.arrayOf(PropTypes.object)
+}
+
