@@ -2,6 +2,7 @@ const bodyParser = require('body-parser');
 const express = require('express');
 
 const rejectIfNoToken = require('../middlewares').rejectIfNoToken;
+const extractTokenFromAuthHeader = require('../helpers/tokenHelper').extractTokenFromAuthHeader;
 
 const gnomesService = require('../services/gnomesService');
 const cartService = require('../services/cartService');
@@ -21,7 +22,7 @@ const getGnomesFromCart = async (cart) => {
 }
 
 const sendGnomeItems = async (token, res) => {
-    const cart = await cartService.getCart(token);    
+    const cart = await cartService.getCart(token);
     if (!cart || !cart.items || cart.items.length === 0) {
         return res.send({ items: [] });
     }
@@ -32,7 +33,8 @@ const sendGnomeItems = async (token, res) => {
 
 router.get('/api/items', rejectIfNoToken, (req, res) => {
     try {
-        sendGnomeItems(req.query.token, res);
+        const token = extractTokenFromAuthHeader(req.header('Authorization'));
+        sendGnomeItems(token, res);
     } catch (error) {
         res.status(500);
         return res.send({ items: [] })
@@ -44,13 +46,14 @@ router.put('/api/items/:item_id', rejectIfNoToken, async (req, res) => {
     console.log('Item targetted %s', req.params.item_id);
 
     try {
-        await cartService.addToCart(req.body.token, req.params.item_id);
+        const token = extractTokenFromAuthHeader(req.header('Authorization'));
+        await cartService.addToCart(token, req.params.item_id);
         const gnomeInCatalog = await gnomesService.getGnome(req.params.item_id);
     
         if(!gnomeInCatalog) {
             await gnomesService.addGnomes([ req.body.item ]);        
         }
-        sendGnomeItems(req.body.token, res);
+        sendGnomeItems(token, res);
 
     } catch (error) {
         res.status(500);
@@ -62,8 +65,9 @@ router.delete('/api/items/:item_id', rejectIfNoToken, async (req, res) => {
     console.log('Item targetted', req.params.item_id);
     
     try {
-        await cartService.removeFromCart(req.body.token, req.params.item_id);
-        sendGnomeItems(req.body.token, res);    
+        const token = extractTokenFromAuthHeader(req.header('Authorization'));
+        await cartService.removeFromCart(token, req.params.item_id);
+        sendGnomeItems(token, res);    
     } catch (error) {
         res.status(500);
         res.send({ items: [] });
